requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
langchain-community==0.0.38
faiss-cpu==1.7.4
transformers==4.35.2
python-json-logger==2.0.7